requirements.txt
grpcio==1.50.0
grpcio_reflection==1.50.0
numpy==1.23.4
protobuf==4.21.9
scikit_learn==1.1.3
scipy==1.9.3
